+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2022 - 2022/01/27
+
+- Update deps
+- Use black to generate help instead of hardcoded
+- Add formal tests
+
 ## 2021.0.1 - 2021/10/12
 
 - Security improvement and unicode support